Move trimNewlines from WordPressKit
